refactor: extract "places" sagas from "device-management" ones
