+++ src/sagas/device-management/index.ts
@@ -1,0 +1,5 @@
+import { sagas as placesSagas } from './places';
+
+export const sagas = [
+  ...placesSagas,
+];
