+++ src/sagas/device-management/places.spec.ts
@@ -1,4 +1,4 @@
-import { fetchDataApi } from './device-management';
+import { fetchDataApi } from './places';
 
 describe('Test1', () => {
   const generator = fetchDataApi({ ids: [1, 2], type: 'FETCH_PLACES' });
