+++ src/sagas/index.ts
@@ -1,9 +1,9 @@
 import { all } from 'redux-saga/effects';
 
-import { dummyDataSagas } from './device-management';
+import { sagas as deviceManagementSagas } from './device-management';
 
 export default function* rootSaga() {
   yield all([
-    ...dummyDataSagas
+    ...deviceManagementSagas
   ]);
 }
